Remove all the custom cloning methods.
