+++ Source/Client/Features/Application/ApplicationState.cs
@@ -1,8 +1,12 @@
 ﻿namespace Client.Features.Application
 {
-  public partial class ApplicationState 
+  using BlazorState;
+
+  public partial class ApplicationState : State<ApplicationState>
   {
     public string Name { get; private set; }
     public string Version => GetType().Assembly.GetName().Version.ToString();
+
+    protected override void Initialize() => Name = "myvault.anthemgold.com";
   }
 }
