+++ Source/Client/Features/Edge/State/EdgeWalletInfo.cs
@@ -3,7 +3,7 @@
   using System;
   using System.Collections.Generic;
 
-  public class EdgeWalletInfo : ICloneable
+  public class EdgeWalletInfo
   {
     public List<string> AppIds { get; set; }
     public bool Archived { get; set; }
@@ -12,12 +12,5 @@
     public Dictionary<string, string> Keys { get; set; }
     public int SortIndex { get; set; }
     public string Type { get; set; }
-
-    public object Clone()
-    {
-      var clone = MemberwiseClone() as EdgeWalletInfo;
-      clone.Keys = new Dictionary<string, string>(Keys);
-      return clone;
-    }
   }
 }
