+++ Test/Client.Integration.Tests/Features/Edge/State/EdgeTransactionUnitTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using AnyClone;
 using Client.Features.Edge.State;
 using Shouldly;
